Add an initial_workers option
This option goes along with `min_workers`, and `max_workers`.  When the
cluster is first brought up (or when it is refreshed with a subsequent
`ray up`) this number of nodes will be started.

It's a workaround for issues of scaling (see related issues) where it
can take a long time (or forever in the case where the head node has
`--num-cpus 0`) to scale up a cluster in response to increasing demand.