Add dependencies for dashboard to installation.rst
Updating the docs to include pip installing `aiohttp` and `psutil`, both of which the dashboard requires.  Since the whole dashboard section is optional, I thought I'd just add it in the docs rather than make it an explicit requirement of the project.  Tell me if you'd prefer them as requirements in the `setup.py`, though.