Add initial implementation of C bindings for Ray
This patch introduces a new directory, `clib`. `clib` includes a
standalone C header (`clib/ray.h`) and builds an associated C-compatible
`libray`. Compiled applications that wish to interact with Ray can use
the C header and link against the generated `libray`. They do not need
to be included in the Ray build pipeline.

`libray` is generated using a thin C++ shim layer (`clib/lib.cpp`) which
presents C function signatures and opaque types and internally makes the
appropriate C++ calls to Ray.

In order to keep `ray.h` self-contained and in pure C, helper methods
are provided for many types. For example, all Ray types that are
`std::vector` are instead `SomeOpaqueType *`, with helper methods like
`type_new`, `type_free`, `type_at`, `type_len`, etc. Similarly, types
like `RayObject` are presented to the user as `Object*`, with helper
functions like `object_buf`, `object_buf_size`, etc.

The bindings are not complete, but cover the main basic building blocks
needed to interact with Ray, including constructing a `CoreWorker`, and
calling the main methods like `Wait`, `Get`, `Put`, and `SubmitTask`.
The API is also overly simplified in places, such as object metadata
always being empty, and various task options not being possible to
specify. This should probably be fixed before an eventual release.

Task dispatch works through a single callback function, whose signature
matches the `TaskExecutionCallback` used internally, but with the types
mapped to their public opaque counterparts. One exception is that only
the _first_ string of the function descriptor is provided as a function
descriptor. This is probably sufficient for most applications, but may
be something we wish to revisit.

Some inefficiencies in the implementation stem from the fact that many
Ray APIs take `std::vector` or `std::shared_ptr`, which are difficult to
construct from C types without copying. Much of this can probably be
optimized.

Oh, yeah, and all the types and functions in ray.h should obviously be
rigorously documented if they are to become the main way for external
programs to interact with Ray.