Add test driver + worker for libray
This patch adds `clib/main.c`, a binary that uses only `ray.h` and
`libray` to implement a proof-of-concept of a compiled Ray client.

The binary assumes that if it is called with `--node-ip-address=`
as its first argument it is being spawned as a worker, otherwise it
assumes the role of a driver for a simple workload that submits a single
task, waits for it, and prints its result. We will probably want a more
reliable mechanism for letting compiled binaries know when they are
being spawned as workers from Ray (as opposed to being invoked directly
by a user to be used as a driver). One option here is to manipulate
`argv[0]`, or to use environment variables.

The demo worker does not implement true function dispatch at the moment.
All tasks end up executing the same code, independent of the argument
that is passed as the function descriptor. It will ultimately be up to
each implementation linking with libray (probably language bindings) to
specify how function descriptors and the dispatch of them work.

The binary does not currently know how to start a Ray cluster. Instead,
it assumes that one is already running, along with all necessary Ray
services. It also assumes that the running Ray cluster was started with
`native_worker_path` pointing to the compiled binary. The easiest way to
do the necessary setup is to run first compile the binary with:

```
$ bazel build //clib:test
```

Then start the cluster from an interactive Python shell with:

```python
addrs = ray.init(native_worker_path='path/to/bazel-bin/clib/test', num_cpus=0)
print(ray.nodes()[0]['NodeManagerPort'], addrs['redis_address'])
```

And then to run the binary as a driver by substituting the values output
above in the following command:

```
$ bazel run //clib:test -- IP_ADDRESS NODE_MANAGER_PORT REDIS_PORT
```

We will probably want a better story for how non-Python drivers can
start a Ray cluster. One good strategy might be to provide a launcher
program that essentially runs `ray.init` for you. That program can then
be run directly from the driver, or automatically by the language
bindings.