Attempt to retry pending requests on actor restart
Previously any actor failure would cause all outstanding actor tasks to
get marked as failed, leading to an exception in the user program.
This patch changes the behavior such that as long as the actor has
remaining reconstructions, all tasks which haven't been completed previously
will get resubmitted to the new instance automatically.

The new semantics allow to handle failures in a more transparent way, but
also requiresthe user to make sure that the actor will always restart in a
state in which the repeated calls are still valid. If this is not the desired
behavior the workaround is to disallow automatic reconstruction and implement
manual handling of actor failures in the user program.