Partially revert "[Core/RLlib] Move `log_once` from rllib to ray.util. (#7273)" (#7361)
This partially reverts commit 357232d1241d20df3d7ec1ee723a26fdbf5cffe9.

The addition of python/__init__.py broke the build on Windows. However, this is difficult to notice because Bazel doesn't seem to notice this dependency. You first have to go to a commit that fails on this issue, and then try to re-build this commit, so that Bazel actually performs a rebuild.

A useful command-line for triggering the exact build i:

bazel build --compile_one_dependency //:python/ray/_raylet.pyx