rsync_down source/target bug.
I supposed that the goal of rsync_down is to copy changes from the cluster back to the source. The current code tries to copy the source folder (not target) from the cluster back to the source folder. It only works if the folder names happen to be the same. This looks like a clear bug to me, but I could be misunderstanding the command.