[docs] Tune Search:  Wrong parameter name
Both on docs and class docstrings, "search_alg=algo" parameter is wrongly used as "algo=algo" on BayesOpt and DragonFly search.

Could not find the corresponding doc files, only docstring of those classes, so this PR only solves the later one.

To find it on docs, search for "algo=algo" in:
https://docs.ray.io/en/latest/tune-searchalg.html#bayesopt-search
and
https://docs.ray.io/en/latest/tune-searchalg.html#dragonfly-search