CoreWorker correctly push metrics to agent
Before this PR, we are not passing in the metrics_agent_port to
each worker. At the startup, each worker will generate a random
port for metric agent and tries to push metrics to that port.
The error is ignored so it was never visible.

This PR fixes this bug by passing metrics_agent_port through the
default_worker.py command arguments.