AWS Spot Fleet support
AWS Spot Fleet is a method of requesting EC2 instances in terms of
resources rather than explicitly specifying individual nodes to launch.
One defines a set of launch specs each of which might correspond with a
difference instance type, along with a so called instance weight, and a
total amount of instance weight to be launched.  AWS then works out the
optimal instance types to launch which would satify the total required
weight.

This PR adds support for spot fleet requests to the autoscaler. CPU
resource in Ray maps to instance weights in the spot fleet request.  To
achieve this a `ResourceLauncher` class has been added alongside the
`NodeLauncher` class which pulls requests to launch CPU resource off a
queue and passes the request onto the node provider.  The AWS node
provider in turn has new methods to handle launching nodes according to
total amount of required CPU (`create_cpu`) and for checking the current
amount of non-terminated CPU (`non_terminated_cpu`).

When operating in fleet mode (ie.
`auto_scaler['config']['fleet']['enabled']` is True), the logic in
`StandardAutoscaler` is changed to use a resource launcher instead of
node launcher and measure desired / running CPUs rather than desired /
running nodes.

Note that when fleet mode is enabled launching nodes via
the node launcher is not possible. In addition, when in fleet mode,
any other resource other than CPU is ignored, since the spot fleet API
only allows one dimension in its instance weighting.