[RLlib] Closes #11924: Add support for custom/ray environments in rollouts.py for agents without workers (#11926)
* Closes #11924
Formerly, rollout.py would only load environments from gym (with
gym.make() ) , if an agent without workers is employed (such as ES or
ARS). This will result in an error, if a custom environment is used. This
PR adds the possibility to load environments from the ray registry,
while maintaining the support for gym environments.

* Update rllib/rollout.py

Co-authored-by: Sven Mika <sven@anyscale.io>