[workflow] Avoid running workflow step multiple times. (#19090)
When workflow recover, it'll try to reconstruct the DAG. However, it's step scoped, which means if a workflow is passed to multiple steps, it'll be executed multiple times which breaks the exactly-once semantic.

For ObjectRef it's ok since it'll be cached with serialization context, but we also need a similar thing for Workflow input.

This logic is put in workflow layer instead of serialization layer because it's dedupe on app layer.

Issue #18997 has race conditions, and it's also related to this one. The reason is that multiple steps will try to issue writes to virtual actors at the same time which is not allowed right now and can lead to race condition.