[gcs] Fix the regression of enabling grpc based broadcasting in actor scheduling (#19664)
## Why are these changes needed?
Previously, we don't send requests if there is an in-flight request. But this is actually bad, because it prevent raylet get the latest information. For example, if the request needs 200ms to arrive at the raylet, the raylet will lose one update. In this case, the next request will arrive after 200 + 100 + (in flight time) ms. So we still should send the request.

TODO:
- Push the snapshot to raylet if the message is lost.
- Handle message loss in raylet better.


## Related issue number
#19438