[ci] Script for building M1 wheels
This PR includes a script for building wheels for Macs with M1 processors. It roughly follows the pattern of the other scripts with a few differences. 

* Manually installs `nvm`
* Uses miniforge conda to install python/pip instead of python foundation .pkgs 
* Doesn't pin numpy (we probably shouldn't be pinning it in the other scripts either...)
* Commit detection falls back to `git` instead of erroring

All of these changes were made so that the script works on a laptop, which comes with a subset of the dependencies that the x86 buildkite image comes with.