[Core] Fail inflight tasks on actor restarting (#19354)
## Why are these changes needed?

If an actor failover is triggered, but the RPC connection between the caller and the crashed actor instance is not disconnected automatically, subsequent tasks to the new actor instance may not be executed. The root cause is that the sequence numbers of tasks sent to the new actor instance is not starting from 0. Details can be found in #14727.

This PR fixes it by ensuring all inflight actor tasks fail immediately when actor failover is detected (via actor state notifications).

## Related issue number

closes #14727