[GCP] Filter GCP TPUs by cluster name, matching behavior for GCP compute nodes. (#20311)
Ray currently does not filter GCP TPU nodes based on the cluster name, resulting in conflicts when multiple ray clusters are running on the same GCP account.

This change updates the TPU behavior to match the GCP compute node behavior, i.e. filtering to TPU nodes for the current cluster.