[hotfix] Fix stress_test_many_tasks cluster environment (#20519)
This should fix the long running release tests that are failing to build their app configs.

It seems like pip install ray[all] now downgrades the ray version. It's unclear why, but most likely, a dependency has pinned the ray version now. This PR explicitely install the version of Ray that we want after the pip install ray[all] to fix the problem.