[Core] Fix worker pool crash due to incorrect `pending_exit_idle_workers_` usage (#20180)
## Why are these changes needed?

When the Java multi-worker feature is on and if workers respond `Exit` requests from the worker pool with delays (even slower than the interval of `TryKillingIdleWorkers`), the worker pool may send additional `Exit` requests to workers before receiving replies of previous ones. This leads to a `RAY_CHECK` failure from here

https://github.com/ray-project/ray/blob/60df705b4e72c5a4a74a282ffce999ed4a30f974/src/ray/raylet/worker_pool.cc#L984

due to executing two reply callbacks in a row.

This PR fixes the bug by ensuring the worker pool only sends new `Exit` requests to a worker if there are no inflight `Exit` requests to any worker of the worker process.