[gcs] Update constructor of gcs client (#21025)
GcsClient accepts only redis before. To make it work without redis, we need to be able to pass gcs address to gcs client as well.

In this PR, we add GCS related into into GcsClientOptions so that we can connect to the gcs directly with gcs address.

This  PR is part of GCS bootstrap. In the following PR, we'll add functionality to set the correct GcsClientOptions based on flags.