[ci] Add repro-ci.py script to automatically setup Buildkite-runner-like instances to debug CI runs (#21292)
Create an AWS instance to reproduce Buildkite CI builds.

This script will take a Buildkite build URL as an argument and create
an AWS instance with the same properties running the same Docker container
as the original Buildkite runner. The user is then attached to this instance
and can reproduce any builds commands as if they were executed within the
runner.

This utility can be used to reproduce and debug build failures that come up
on the Buildkite runner instances but not on a local machine.